+++ profiles/views/globalscope.py
@@ -1,18 +1,18 @@
 from Squest.utils.squest_views import *
-from profiles.forms import GlobalPermissionForm
-from profiles.models import GlobalPermission
+from profiles.forms import GlobalScopeForm
+from profiles.models import GlobalScope
 from profiles.tables import UserRoleTable, PermissionTable
 
 
-class GlobalPermissionRBACView(SquestDetailView):
-    model = GlobalPermission
-    template_name = "profiles/globalpermission_rbac_detail.html"
+class GlobalScopeRBACView(SquestDetailView):
+    model = GlobalScope
+    template_name = "profiles/globalscope_rbac_detail.html"
 
     def get_permission_required(self):
-        return "profiles.view_users_globalpermission"
+        return "profiles.view_users_globalscope"
 
     def get_object(self, queryset=None):
-        return GlobalPermission.load()
+        return GlobalScope.load()
 
     def dispatch(self, request, *args, **kwargs):
         self.kwargs['pk'] = self.get_object().id
@@ -22,17 +22,17 @@ class GlobalPermissionRBACView(SquestDetailView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context['breadcrumbs'] = None
-        context['title'] = "Global permission"
+        context['title'] = "Global scope"
         context['users'] = UserRoleTable(self.object.users)
         return context
 
 
-class GlobalPermissionDefaultPermissionView(SquestDetailView):
-    model = GlobalPermission
-    template_name = "profiles/global_permission_default_permission_detail.html"
+class GlobalScopeDefaultPermissionView(SquestDetailView):
+    model = GlobalScope
+    template_name = "profiles/global_scope_default_permission_detail.html"
 
     def get_object(self, queryset=None):
-        return GlobalPermission.load()
+        return GlobalScope.load()
 
     def dispatch(self, request, *args, **kwargs):
         self.kwargs['pk'] = self.get_object().id
@@ -43,21 +43,24 @@ class GlobalPermissionDefaultPermissionView(SquestDetailView):
         context = super().get_context_data(**kwargs)
         context['breadcrumbs'] = None
         context['title'] = "Default permissions"
-        permission_table = PermissionTable(self.object.default_permissions.all())
+        permission_table = PermissionTable(self.object.global_permissions.all())
         permission_table.exclude = ("actions",)
-        context['default_permissions'] = permission_table
+        context['global_permissions'] = permission_table
+        owner_permission_table = PermissionTable(self.object.owner_permissions.all())
+        owner_permission_table.exclude = ("actions",)
+        context['owner_permissions'] = owner_permission_table
         return context
 
 
-class GlobalPermissionEditView(SquestUpdateView):
-    model = GlobalPermission
-    form_class = GlobalPermissionForm
+class GlobalScopeEditView(SquestUpdateView):
+    model = GlobalScope
+    form_class = GlobalScopeForm
 
     def get_object(self, queryset=None):
-        return GlobalPermission.load()
+        return GlobalScope.load()
 
     def get_success_url(self):
-        return reverse_lazy('profiles:globalpermission_default_permissions')
+        return reverse_lazy('profiles:globalscope_default_permissions')
 
     def dispatch(self, request, *args, **kwargs):
         self.kwargs['pk'] = self.get_object().id
@@ -67,7 +70,7 @@ class GlobalPermissionEditView(SquestUpdateView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context['breadcrumbs'] = [
-            {'text': 'Global permission', 'url': reverse('profiles:globalpermission_rbac')},
+            {'text': 'Global scope', 'url': reverse('profiles:globalscope_rbac')},
             {'text': f'Edit', 'url': ""},
         ]
         return context