+++ service_catalog/models/support.py
@@ -52,9 +52,22 @@ class Support(SquestModel):
 
     @classmethod
     def get_q_filter(cls, user, perm):
+        app_label, codename = perm.split(".")
+        from profiles.models import GlobalScope
+        globalscope = GlobalScope.load()
+        additional_q = Q()
+        if globalscope.owner_permissions.filter(
+                codename=codename,
+                content_type__app_label=app_label
+        ).exists():
+            additional_q = Q(opened_by=user)
+
         return Q(
            instance__in=Instance.get_queryset_for_user(user, perm)
-        )
+        ) | additional_q
+
+    def is_owner(self, user):
+        return self.instance.is_owner(user) or self.opened_by == user
 
     def get_scopes(self):
         return self.instance.get_scopes()