+++ tests/test_profiles/test_model/test_get_queryset_instance.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.models import User
 
-from profiles.models import Team, Organization, GlobalPermission, Role
+from profiles.models import Team, Organization, GlobalScope, Role
 from profiles.models.squest_permission import Permission
 
 from service_catalog.models import Instance
@@ -27,7 +27,7 @@ class TestModelScopeGetQuerysetInstance(TransactionTestUtils):
                                                         content_type__model="instance",
                                                         codename=codename)
         self.role_view_instance.permissions.add(self.permission_object)
-        self.global_perm = GlobalPermission.load()
+        self.global_scope = GlobalScope.load()
 
     def _assert_can_see_everything(self, user):
         self.assertQuerysetEqualID(Instance.get_queryset_for_user(user, self.permission),
@@ -37,9 +37,34 @@ class TestModelScopeGetQuerysetInstance(TransactionTestUtils):
         self.assertQuerysetEqualID(Instance.get_queryset_for_user(user, self.permission),
                                    Instance.objects.none())
 
-    def test_get_queryset_globalpermission_default_permissions(self):
+    def test_get_queryset_when_owner(self):
+        # No instances
+        self._assert_can_see_nothing(self.superuser)
+        self._assert_can_see_nothing(self.user1)
+
+        # create a new instance
+        instance1 = Instance.objects.create(name="Instance #1", quota_scope=self.default_quota_scope,
+                                            requester=self.user1)
+        self.assertTrue(instance1 in Instance.objects.all())
+        self.assertEqual(Instance.objects.count(), 1)
+
+        # add global perm
+        self.global_scope.owner_permissions.add(self.permission_object)
+
+        # everyone can see
+        self._assert_can_see_everything(self.superuser)
+        self._assert_can_see_everything(self.user1)
+
+        # remove global perm
+        self.global_scope.owner_permissions.remove(self.permission_object)
+
+        # only super can see
+        self._assert_can_see_everything(self.superuser)
+        self._assert_can_see_nothing(self.user1)
+
+    def test_get_queryset_globalscope_global_permissions(self):
         """
-        Test the default_permissions in global permission
+        Test the global_permissions in global scope
         """
 
         # No instances
@@ -52,22 +77,22 @@ class TestModelScopeGetQuerysetInstance(TransactionTestUtils):
         self.assertEqual(Instance.objects.count(), 1)
 
         # add global perm
-        self.global_perm.default_permissions.add(self.permission_object)
+        self.global_scope.global_permissions.add(self.permission_object)
 
         # everyone can see
         self._assert_can_see_everything(self.superuser)
         self._assert_can_see_everything(self.user1)
 
         # remove global perm
-        self.global_perm.default_permissions.remove(self.permission_object)
+        self.global_scope.global_permissions.remove(self.permission_object)
 
         # only super can see
         self._assert_can_see_everything(self.superuser)
         self._assert_can_see_nothing(self.user1)
 
-    def test_get_queryset_globalpermission_perm_specific_user(self):
+    def test_get_queryset_globalscope_perm_specific_user(self):
         """
-        Test the rbac role in global permission
+        Test the rbac role in global scope
         """
         # No instances
         self._assert_can_see_nothing(self.superuser)
@@ -80,7 +105,7 @@ class TestModelScopeGetQuerysetInstance(TransactionTestUtils):
         self.assertEqual(Instance.objects.count(), 1)
 
         # assign a view instance to user1
-        self.global_perm.add_user_in_role(self.user1, self.role_view_instance)
+        self.global_scope.add_user_in_role(self.user1, self.role_view_instance)
 
         # everyone can see except user2
         self._assert_can_see_everything(self.superuser)
@@ -88,13 +113,12 @@ class TestModelScopeGetQuerysetInstance(TransactionTestUtils):
         self._assert_can_see_nothing(self.user2)
 
         # unassign view instance role to user1
-        self.global_perm.remove_user_in_role(self.user1, self.role_view_instance)
+        self.global_scope.remove_user_in_role(self.user1, self.role_view_instance)
 
         # only super can see
         self._assert_can_see_everything(self.superuser)
         self._assert_can_see_nothing(self.user1)
         self._assert_can_see_nothing(self.user2)
-
 
     ####################
     def test_get_queryset_instance_with_organization_role_with_quota_scope(self):