+++ tests/test_profiles/test_model/test_get_queryset_scope_consume_quota.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.models import User
 
-from profiles.models import Team, Organization, GlobalPermission, Role, Scope
+from profiles.models import Team, Organization, GlobalScope, Role, Scope
 from profiles.models.squest_permission import Permission
 
 from tests.utils import TransactionTestUtils
@@ -15,7 +15,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         self.user3 = User.objects.create_user('user3', 'user3@hpe.com', "password")
         self.superuser = User.objects.create_superuser("superuser")
 
-        self.global_perm = GlobalPermission.load()
+        self.global_scope = GlobalScope.load()
         self.empty_role = Role.objects.create(name="Empty role")
         self.permission_object = Permission.objects.get(content_type__app_label="profiles",
                                                         content_type__model="scope",
@@ -29,9 +29,9 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         perm = "profiles.consume_quota_scope"
         self.assertQuerysetEqualID(Scope.get_queryset_for_user(user, perm), Scope.objects.none())
 
-    def test_get_queryset_organization_by_querying_scope_with_global_permission_role(self):
+    def test_get_queryset_organization_by_querying_scope_with_global_scope_role(self):
         """
-        Test that we can consume all quota by querying Scope if permission is given in a GlobalPermission's role
+        Test that we can consume all quota by querying Scope if permission is given in a GlobalScope's role
         """
 
         # Only superuser can see
@@ -44,7 +44,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         role.permissions.add(self.permission_object)
 
         # Assign role to view Scope to user1
-        self.global_perm.add_user_in_role(self.user1, role)
+        self.global_scope.add_user_in_role(self.user1, role)
 
         # Create a new organization
         org1 = Organization.objects.create(name="Scope #1")
@@ -57,16 +57,16 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         self._assert_can_consume_only_quota_qs(self.superuser, quota_qs=Scope.objects.filter(id__in=[org1.id]))
 
         # Remove permission to user1
-        self.global_perm.remove_user_in_role(self.user1, role)
+        self.global_scope.remove_user_in_role(self.user1, role)
 
         # Only superuser can see
         self._assert_cannot_consume_quota(self.user1)
         self._assert_cannot_consume_quota(self.user2)
         self._assert_can_consume_only_quota_qs(self.superuser, quota_qs=Scope.objects.filter(id__in=[org1.id]))
 
-    def test_get_queryset_organization_by_querying_scope_with_global_permission__default_permissions(self):
+    def test_get_queryset_organization_by_querying_scope_with_global_scope__global_permissions(self):
         """
-        Test that we can consume Organization quota by querying Scope if permission is given in a GlobalPermission's perm
+        Test that we can consume Organization quota by querying Scope if permission is given in a GlobalScope's perm
         """
 
         # Only superuser can see
@@ -77,7 +77,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         # Create Permission/Role_scope")
 
         # Assign permission to all users
-        self.global_perm.default_permissions.add(self.permission_object)
+        self.global_scope.global_permissions.add(self.permission_object)
 
         # Create a new organization
         org1 = Organization.objects.create(name="Organization #1")
@@ -90,16 +90,16 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         self._assert_can_consume_only_quota_qs(self.superuser, quota_qs=Scope.objects.filter(id__in=[org1.id]))
 
         # Remove permission to all users
-        self.global_perm.default_permissions.remove(self.permission_object)
+        self.global_scope.global_permissions.remove(self.permission_object)
 
         # Only superuser can see
         self._assert_cannot_consume_quota(self.user1)
         self._assert_cannot_consume_quota(self.user2)
         self._assert_can_consume_only_quota_qs(self.superuser, quota_qs=Scope.objects.filter(id__in=[org1.id]))
 
-    def test_get_queryset_team_by_querying_scope_with_global_permission_role(self):
+    def test_get_queryset_team_by_querying_scope_with_global_scope_role(self):
         """
-        Test that we can consume Team quota by querying Scope if permission is given in a GlobalPermission's role
+        Test that we can consume Team quota by querying Scope if permission is given in a GlobalScope's role
         """
         # Only superuser can see
         self._assert_cannot_consume_quota(self.user1)
@@ -111,7 +111,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         role.permissions.add(self.permission_object)
 
         # Assign role to user1
-        self.global_perm.add_user_in_role(self.user1, role)
+        self.global_scope.add_user_in_role(self.user1, role)
 
         # Create a new organization/team
         org1 = Organization.objects.create(name="Organization #1")
@@ -126,7 +126,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
                                                quota_qs=Scope.objects.filter(id__in=[org1.id, team1.id]))
 
         # Remove role to user1
-        self.global_perm.remove_user_in_role(self.user1, role)
+        self.global_scope.remove_user_in_role(self.user1, role)
 
         # Only superuser can see
         self._assert_cannot_consume_quota(self.user1)
@@ -134,9 +134,9 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         self._assert_can_consume_only_quota_qs(self.superuser,
                                                quota_qs=Scope.objects.filter(id__in=[org1.id, team1.id]))
 
-    def test_get_queryset_team_by_querying_scope_with_global_permission__default_permissions(self):
+    def test_get_queryset_team_by_querying_scope_with_global_scope__global_permissions(self):
         """
-        Test that we can consume Team quota if permission is given in a GlobalPermission's perm
+        Test that we can consume Team quota if permission is given in a GlobalScope's perm
         """
         # Only superuser can see
         self._assert_cannot_consume_quota(self.user1)
@@ -144,7 +144,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
         self._assert_cannot_consume_quota(self.superuser)
 
         # Create Permission/Role
-        self.global_perm.default_permissions.add(self.permission_object)
+        self.global_scope.global_permissions.add(self.permission_object)
 
         # Create a new organization/team
         org1 = Organization.objects.create(name="Organization #1")
@@ -159,7 +159,7 @@ class TestModelScopeGetQuerysetConsumeQuota(TransactionTestUtils):
                                                quota_qs=Scope.objects.filter(id__in=[org1.id, team1.id]))
 
         # Remove role to everyone
-        self.global_perm.default_permissions.remove(self.permission_object)
+        self.global_scope.global_permissions.remove(self.permission_object)
 
         # Only superuser can see
         self._assert_cannot_consume_quota(self.user1)