+++ tests/test_profiles/test_model/test_who_has_perm_instance.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.models import User
 
-from profiles.models import Organization, GlobalPermission, Role, Team
+from profiles.models import Organization, GlobalScope, Role, Team
 from profiles.models.squest_permission import Permission
 from service_catalog.models import Instance
 
@@ -12,7 +12,7 @@ class TestModelWhoHasPermOnModelLinkedToInstances(TransactionTestUtils):
     def setUp(self):
         super(TestModelWhoHasPermOnModelLinkedToInstances, self).setUp()
 
-        self.global_perm = GlobalPermission.load()
+        self.global_scope = GlobalScope.load()
         self.empty_role = Role.objects.create(name="Empty role")
 
         self.test_org = Organization.objects.create(name="Org")
@@ -38,14 +38,14 @@ class TestModelWhoHasPermOnModelLinkedToInstances(TransactionTestUtils):
         self.assertQuerysetEqualID(user_with_permissions, User.objects.filter(is_superuser=True))
         self.assertNotIn(self.user_with_no_perm, user_with_permissions)
 
-    def test_who_has_perm_on_instance_globalperm_default_permissions(self):
+    def test_who_has_perm_on_instance_global_scope_global_permissions(self):
         # user_with_permissions is empty
         user_with_permissions = self.instance1.who_has_perm('service_catalog.view_instance')
         self.assertQuerysetEqualID(user_with_permissions, User.objects.filter(is_superuser=True))
         self.assertNotIn(self.user_with_no_perm, user_with_permissions)
 
-        # Permission given in GlobalPerm as default_permissions
-        self.global_perm.default_permissions.add(self.view_instance_permission)
+        # Permission given in GlobalScope as global_permissions
+        self.global_scope.global_permissions.add(self.view_instance_permission)
 
         # Permissions given to all Squest user
         user_with_permissions = self.instance1.who_has_perm('service_catalog.view_instance')
@@ -130,15 +130,15 @@ class TestModelWhoHasPermOnModelLinkedToInstances(TransactionTestUtils):
         self.assertIn(self.user, user_with_permissions)
         self.assertNotIn(self.user_with_no_perm, user_with_permissions)
 
-    def test_who_has_perm_on_instance_globalperm_rbac(self):
+    def test_who_has_perm_on_instance_global_scope_rbac(self):
         # Test Team RBAC
         ## user is not in Team
         user_with_permissions = self.instance1.who_has_perm('service_catalog.view_instance')
         self.assertNotIn(self.user, user_with_permissions)
         self.assertNotIn(self.user_with_no_perm, user_with_permissions)
 
-        ## Add view_instance_role in globalperm for self.user
-        self.global_perm.add_user_in_role(self.user, self.view_instance_role)
+        ## Add view_instance_role in global_scope for self.user
+        self.global_scope.add_user_in_role(self.user, self.view_instance_role)
 
         ## user is now in the list
         user_with_permissions = self.instance1.who_has_perm('service_catalog.view_instance')